@@ -5,7 +5,7 @@ version: 0.3.11
 maintainer: Rafal Wachol (@RafalWachol) 
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 
 flutter:
   plugin:
@@ -17,6 +17,6 @@ dependencies:
     sdk: flutter
 
 dev_dependencies:
-  mockito: 
+  mockito: ^5.0.17 
   flutter_test:
     sdk: flutter
